function calculateWinner changed to folder
