+++ react/tic-tac-toe/src/app/screens/Game/index.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
 
 import { PLAYER_TYPE } from '../../../constants/player';
+import { calculateWinner } from '../../../utils/calculateWinner';
 
 import Board from './components/Board';
 import styles from './styles.module.scss';
+
 
 class Game extends Component {
   state = {
@@ -21,35 +23,13 @@ class Game extends Component {
       return;
     }
     squares[i] = xIsNext ? PLAYER_TYPE.X : PLAYER_TYPE.O;
-    const winner = this.calculateWinner(squares);
+    const winner = calculateWinner(squares);
     this.setState({
       history: history.concat([{ squares }]),
       xIsNext: !xIsNext,
       stepNumber: history.length,
       winner
     });
-  }
-
-  calculateWinner = (squares) => {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6]
-    ];
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      } else if (!squares.includes(null)) {
-        return PLAYER_TYPE.TIE;
-      }
-    }
-    return null;
   }
 
   jumpTo = step => () => {
