+++ react/tic-tac-toe/src/services/MatchesService.js
@@ -1,7 +1,11 @@
+import { wrapService } from 'redux-recompose';
+
 import api from '../config/api';
 
-export default {
+const service = {
   getMatches: () => new Promise(resolve => setTimeout(() => api.get('/matches').then(response => resolve(response)), 1000)),
   onLogin: values => api.post('/login', values),
   createMatches: values => api.post('/matches', values)
 };
+
+export default wrapService(service, 'soccer');
